*Functioning version of the processing main script
The script runs denoising, GNLC of AFI and MESE images, and B1correction/T2fitting as sequential jobs. After that, the working directory of each session and then the the remains of the working directory is cleaned up in separate jobs. The dependency of the SLURM jobs is currently handled by intermediate/bridge jobs that submit the actual processing jobs and handle the job dependencies.