*Actual python processing scripts using PyMRItools (adapted from Jupyter Notebooks) + helper script
One processing script handles the denoising of the MESE data. The other one does the B1+ correction and the T2 fitting via pattern matching. In between, the Gradient nonlinearity correction has to be performed. The python scripts also transfer and/or adapt the NIfTI headers and copy the corresponding json files. // The helper scripts provide functions that are used regularly across the scripts (e.g., for saving, loading, copying jsons, plotting). The plotting functions are not very generalized, and rather put into this separate script to unclutter the processing python scripts than for re-use purposes.