*Bugfix: File exist checks before submitting T2 processing jobs
Submitting jobs for sessions that exist but don't contain MESE and AFI data led to zombie jobs that are queued but never satisfied their dependencies due to cancellation of the first job in the pipeline. This issue is now avoided by assuring that the required data exists before submitting the jobs. Zombie tasks can still arise but will now hint to an actual issue in the processing.